Renamed test folders.
spatial decomposition updated Pythonic versions are in test.ipynb.

We are having difficulty getting the Python and MATLAB outputs for spatial decomposition to agree. Eric and I believe that exact agreement is not necessary as long as the eventual outputs (to whatever end spatial decomposition is used for, idk) are correct.

Lastly, Eric was able to figure out the correct implementation of the Householder matrix initialization. The problem was order of operations with `np.matmult`.

So, in summation, we still need to test spatial_decomposition and its sub-functions. The spatial_decomp_test folder (or whatever it's called) should contain all files necessary to run the Python and MATLAB code in parallel to find discrepancies and debug.

It was a long but rewarding hackathon. I'm looking forward to finishing the work I started on this library at some point. Thanks to Austin Gerlt (@argerlt) for setting this up.

Rohan